requirements/local.txt
-r base.txt
beautifulsoup4==4.5.1
django-webtest==1.8.0
six==1.10.0
waitress==1.0.1
WebOb==1.6.2
WebTest==2.0.23
coverage==4.2
django-discover-runner==1.0
django-debug-toolbar==1.6
factory-boy==2.7.0
fake-factory==0.7.2
python-dateutil==2.5.3


